--- src/domain/community/virtualContributor/components/BasicSpaceCard.tsx
@@ -1,9 +1,0 @@
-// TODO: add cardBanner if we want support of Spaces as BOK
-export interface BasicSpaceProps {
-  avatar?: {
-    uri: string;
-  };
-  displayName: string;
-  tagline?: string;
-  url: string;
-}
